requirements.txt
cassandra-driver==3.14.0
certifi==2020.6.20
click==7.1.2
Flask==1.1.2
geomet==0.2.1.post1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
opencv-python==4.4.0.40
Pillow==7.2.0
pip==20.2.2
setuptools==49.6.0.post20200814
six==1.15.0
Werkzeug==1.0.1
wheel==0.34.2
wincertstore==0.2